cleaned original template project * removed sample rest endpoints * renamed root package to tc (technical challenge)
