+++ src/main/groovy/com/github/mrduguo/tc/Application.groovy
@@ -1,0 +1,15 @@
+package com.github.mrduguo.tc
+
+import org.springframework.boot.SpringApplication
+import org.springframework.boot.autoconfigure.SpringBootApplication
+
+@SpringBootApplication
+public class Application {
+
+
+	public static void main(String[] args) {
+		System.properties['info.app.start.time']=new Date().format('yyyy-MM-dd HH:mm:ss z')
+		SpringApplication.run(Application.class, args)
+	}
+
+}
