--- src/test/groovy/com/github/mrduguo/gradle/sampleapp/BeanSpec.groovy
@@ -1,16 +1,0 @@
-package com.github.mrduguo.gradle.sampleapp
-
-import org.springframework.beans.factory.annotation.Autowired
-
-class BeanSpec extends AbstractSpec {
-
-    @Autowired
-    Application application
-
-
-    void "should return Greetings from application"() {
-        expect:
-        application.index() == 'Hello My World'
-    }
-
-}
